plugin updates

+++ wp-content/plugins/jetpack/uninstall.php
@@ -5,7 +5,8 @@
  * @package automattic/jetpack
  */
 
-use Automattic\Jetpack\Backup\Helper_Script_Manager;
+use Automattic\Jetpack\Backup\V0005\Helper_Script_Manager;
+use Automattic\Jetpack\Connection\Manager as Connection_Manager;
 use Automattic\Jetpack\Sync\Sender;
 
 /**
@@ -18,7 +19,7 @@ function jetpack_uninstall() {
 		dirname( WP_UNINSTALL_PLUGIN ) !== dirname( plugin_basename( __FILE__ ) )
 	) {
 		status_header( 404 );
-		exit;
+		exit( 0 );
 	}
 
 	if ( ! defined( 'JETPACK__PLUGIN_DIR' ) ) {
@@ -26,6 +27,11 @@ function jetpack_uninstall() {
 	}
 
 	require JETPACK__PLUGIN_DIR . 'vendor/autoload_packages.php';
+
+	if ( method_exists( Connection_Manager::class, 'is_ready_for_cleanup' ) && ! Connection_Manager::is_ready_for_cleanup( dirname( plugin_basename( __FILE__ ) ) ) ) {
+		// There are other active Jetpack plugins, no need for cleanup.
+		return;
+	}
 
 	Jetpack_Options::delete_all_known_options();
